created dispatcher
created the dispatcher, but now I'm thinking I may not need it. I may be getting ahead of myself. We do not know if the monthly user base will be so active that I need to worry about concurrency. Simply implementing a waiting queue at the app.py level may suffice because file sizes may not be large enough that processing takes that long.